project/requirements.txt
# Audio transcription
deepgram_sdk==4.8.1

# LLM APIs
groq==0.31.0
httpx==0.28.1

# LangChain components
langchain_core==0.3.72
langchain_groq==0.3.7
langchain_openai==0.3.28  # Requires openai>=1.86.0
langgraph==0.6.3

# Data processing
numpy==1.26.4  # Compatible with Python 3.10
openai==1.86.0  # Must be >=1.86.0 for langchain_openai compatibility
pandas==2.0.3  # Compatible with Python 3.10
pymongo==4.14.0
python-dotenv==1.1.1

# Machine learning
scikit_learn==1.3.2  # Compatible with Python 3.10

# UI
streamlit==1.48.0
streamlit_feedback==0.1.4

# Twitter API 
twikit==2.3.3

# Vector database and embeddings
lancedb==0.24.2
sentence-transformers==2.2.2  # Note: package name uses hyphen, not underscore
tiktoken==0.10.0
